New taskPrompt in utils/prompt
* moves new taskPrompt to utils/prompt
* updates module exports in existing commands.
