+++ src/cmd/archiveList.js
@@ -2,7 +2,8 @@
 
 const log = require('../utils/log.js');
 const config = require('../utils/config.js');
-const prompt = require('../utils/prompt.js');
+const { prompt } = require('../utils/prompt.js');
+
 const finish = require('../utils/finish.js');
