+++ src/utils/prompt.js
@@ -2,13 +2,13 @@
 
 const finish = require('./finish.js');
 const log = require('./log.js');
-
+const promptSync = require('prompt-sync')();
 
 /**
  * Prompt the User for a set of answers
  * @param {string} prompt One or more prompts to question the User
  * @param {function} callback Callback function(answers)
- * @deprecated Use taskPrompt from newPrompt
+ * @deprecated Use taskPrompt instead
  */
 function prompt(prompt, callback) {
 
@@ -97,5 +97,29 @@ function _finish(answers, promptcount, callback) {
 }
 
 
+/**
+ * Prompts user for integer input until they enter a blank line
+ * @param {string} text 
+ * @returns {number[]} array of user input integers
+ */
+ function taskPrompt(text) {
+  let tasks = []
+  let n
+  log.style('[Enter a blank line when finished]', 'dim', true);
+  while ( n != '' ) {
+      n = promptSync(text)
+      if (parseInt(n) ) {
+          tasks.push(n)
+       } else if (n == '') {
+          break
+       } else {
+          log.style('Enter a task index number, please.','red',true)
+       } 
+  }
 
-module.exports = prompt;
+  return tasks
+
+}
+
+
+module.exports = {prompt,taskPrompt};
